Split Events into new file
